% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ent.R
\name{calc_ent}
\alias{calc_ent}
\title{Calculate the Enthalpy Change of a Reaction}
\usage{
calc_ent(prod, reac, use_hdata = TRUE)
}
\arguments{
\item{prod}{A character vector of product names or formulas, or a numeric vector of product enthalpy values.}

\item{reac}{A character vector of reactant names or formulas, or a numeric vector of reactant enthalpy values.}

\item{use_hdata}{A logical indicating whether to use the hdata dataset (default is TRUE).}
}
\value{
The enthalpy change of the reaction.
}
\description{
This function calculates the enthalpy change (ΔH) for a given reaction
using either user-provided enthalpy values or those from the hdata dataset.
Users can provide either product/reactant names or chemical formulas.
}
\examples{
# Using custom values
calc_ent(c(-285.83, -393.51), c(-74.87, -45.9), use_hdata = FALSE)

calc_ent(prod = c(-285.83, -393.51), reac = c(-74.87, -45.9), use_hdata = FALSE)

prod <- c(-285.83, -393.51)
reac <- c(-74.87, -45.9)
calc_ent(prod, reac, use_hdata = FALSE)

# Using the hdata dataset with names
calc_ent(c("Water", "Carbon dioxide"), c("Methane", "Ammonia"))

calc_ent(prod = c("Water", "Carbon dioxide"), reac = c("Methane", "Ammonia"))

prod <- c("Water", "Carbon dioxide")
reac <- c("Methane", "Ammonia")
calc_ent(prod, reac)

# Using the hdata dataset with formulas
calc_ent(c("H2O", "CO2"), c("CH4", "NH3"))

calc_ent(prod = c("H2O", "CO2"), reac = c("CH4", "NH3"))

prod <- c("H2O", "CO2")
reac <- c("CH4", "NH3")
calc_ent(prod, reac)
}
