% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-calc.R
\name{energy_calc}
\alias{energy_calc}
\title{Calculate Total Energy Consumption}
\usage{
energy_calc(power, duration, use_hdata = TRUE)
}
\arguments{
\item{power}{A numeric vector of power consumed in watts (W) or kilowatts (kW).}

\item{duration}{A numeric vector of duration in hours (h).}

\item{use_hdata}{A boolean indicating whether to use the hdata dataset (default is TRUE).}
}
\value{
Total energy consumption and a message about energy reduction.
}
\description{
Calculate Total Energy Consumption

This function calculates the total energy consumption from the power consumed
and duration, allowing the user to use their own data or a dataset.
}
\examples{
# Using hdata
energy_calc(power, duration)

# Using custom values
power <- c(1000, 1500, 2000)  # in watts
duration <- c(2, 1.5, 3)  # in hours
energy_calc(power, duration, use_hdata = FALSE)
}
\keyword{internal}
