% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity-calc.R
\name{intensity_calc}
\alias{intensity_calc}
\title{Calculate Light Intensity}
\usage{
intensity_calc(intensity, production, use_hdata = TRUE)
}
\arguments{
\item{intensity}{A numeric vector representing the measured light intensity values (in watts).}

\item{production}{A numeric vector representing the corresponding production values (in watts).}

\item{use_hdata}{A boolean indicating whether to use the predefined dataset from EcoTherm (default is TRUE).}
}
\value{
A message indicating the total light intensity and recommendations for intensity reduction if necessary.
}
\description{
Calculate Light Intensity

This function calculates the light intensity based on measured intensity and production values,
allowing the user to use their own data or a predefined dataset.
}
\examples{
# Using hdata
intensity_calc(intensity, production)

# Using custom values
custom_intensity <- c(1000, 1200, 1100)
custom_production <- c(500, 600, 700)
intensity_calc(custom_intensity, custom_production, use_hdata = FALSE)
}
\keyword{internal}
